+++ flutter/lib/data/repositories/queue_repository_impl.dart
@@ -129,6 +129,7 @@ class QueueRepositoryImpl implements QueueRepository {
           createdAtMillis: p.createdAt,
           userName: u.name,
           userPhone: u.phone,
+          userNationalId: u.nationalId,
           storeName: s.name,
         );
       }).toList();
@@ -160,6 +161,7 @@ class QueueRepositoryImpl implements QueueRepository {
         createdAtMillis: p.createdAt,
         userName: u.name,
         userPhone: u.phone,
+        userNationalId: u.nationalId,
         storeName: s.name,
       );
     }).toList();
@@ -197,6 +199,7 @@ class QueueRepositoryImpl implements QueueRepository {
       createdAtMillis: p.createdAt,
       userName: u.name,
       userPhone: u.phone,
+      userNationalId: u.nationalId,
       storeName: s.name,
     );
   }
@@ -226,6 +229,7 @@ class QueueRepositoryImpl implements QueueRepository {
       createdAtMillis: p.createdAt,
       userName: u.name,
       userPhone: u.phone,
+      userNationalId: u.nationalId,
       storeName: s.name,
     );
   }
@@ -254,6 +258,7 @@ class QueueRepositoryImpl implements QueueRepository {
         createdAtMillis: p.createdAt,
         userName: u.name,
         userPhone: u.phone,
+        userNationalId: u.nationalId,
         storeName: s.name,
       );
     });
@@ -315,6 +320,7 @@ class QueueRepositoryImpl implements QueueRepository {
         createdAtMillis: now,
         userName: user.name,
         userPhone: user.phone,
+        userNationalId: user.nationalId,
         storeName: store.name,
       );
     });